+++ docs/html/search/all_12.js
@@ -1,7 +1,7 @@
 var searchData=
 [
-  ['table_20of_20contents_0',['Table of contents',['../index.html#main_table_of_contents',1,'']]],
-  ['thread_20safety_1',['Thread safety',['../general_considerations.html#general_considerations_thread_safety',1,'']]],
+  ['thread_20safety_0',['Thread safety',['../general_considerations.html#general_considerations_thread_safety',1,'']]],
+  ['to_20use_20custom_20pools_1',['When not to use custom pools',['../custom_memory_pools.html#custom_memory_pools_when_not_use',1,'']]],
   ['to_20vulkan_20functions_2',['Pointers to Vulkan functions',['../configuration.html#config_Vulkan_functions',1,'']]],
   ['total_3',['total',['../struct_vma_total_statistics.html#a76f1935f7101883f5bb2a03b6c5649d2',1,'VmaTotalStatistics']]],
   ['type_4',['Choosing memory type',['../choosing_memory_type.html',1,'index']]],