+++ docs/html/search/all_13.js
@@ -6,8 +6,12 @@ var searchData=
   ['unusedrangesizemin_3',['unusedRangeSizeMin',['../struct_vma_detailed_statistics.html#a830eda847ed735d0e91da25cfcf797a4',1,'VmaDetailedStatistics']]],
   ['upload_4',['Staging copy for upload',['../usage_patterns.html#usage_patterns_staging_copy_upload',1,'']]],
   ['uploading_5',['Advanced data uploading',['../usage_patterns.html#usage_patterns_advanced_data_uploading',1,'']]],
-  ['usage_6',['usage',['../staying_within_budget.html#staying_within_budget_controlling_memory_usage',1,'Controlling memory usage'],['../debugging_memory_usage.html',1,'Debugging incorrect memory usage'],['../struct_vma_budget.html#a84dd1ecca8b0110259eb206dbadb11f6',1,'VmaBudget::usage'],['../struct_vma_allocation_create_info.html#accb8b06b1f677d858cb9af20705fa910',1,'VmaAllocationCreateInfo::usage'],['../choosing_memory_type.html#choosing_memory_type_usage',1,'Usage'],['../vk_ext_memory_priority.html#vk_ext_memory_priority_usage',1,'Usage'],['../vk_amd_device_coherent_memory.html#vk_amd_device_coherent_memory_usage',1,'Usage'],['../enabling_buffer_device_address.html#enabling_buffer_device_address_usage',1,'Usage']]],
-  ['usage_20patterns_7',['Recommended usage patterns',['../usage_patterns.html',1,'index']]],
-  ['use_20cases_8',['Other use cases',['../usage_patterns.html#usage_patterns_other_use_cases',1,'']]],
-  ['user_20data_9',['user data',['../allocation_annotation.html',1,'Allocation names and user data'],['../allocation_annotation.html#allocation_user_data',1,'Allocation user data']]]
+  ['usage_6',['Controlling memory usage',['../staying_within_budget.html#staying_within_budget_controlling_memory_usage',1,'']]],
+  ['usage_7',['usage',['../debugging_memory_usage.html',1,'Debugging incorrect memory usage'],['../struct_vma_budget.html#a84dd1ecca8b0110259eb206dbadb11f6',1,'VmaBudget::usage'],['../struct_vma_allocation_create_info.html#accb8b06b1f677d858cb9af20705fa910',1,'VmaAllocationCreateInfo::usage']]],
+  ['usage_8',['Usage',['../choosing_memory_type.html#choosing_memory_type_usage',1,'Usage'],['../vk_ext_memory_priority.html#vk_ext_memory_priority_usage',1,'Usage'],['../vk_amd_device_coherent_memory.html#vk_amd_device_coherent_memory_usage',1,'Usage'],['../enabling_buffer_device_address.html#enabling_buffer_device_address_usage',1,'Usage']]],
+  ['usage_20patterns_9',['Recommended usage patterns',['../usage_patterns.html',1,'index']]],
+  ['use_20cases_10',['Other use cases',['../usage_patterns.html#usage_patterns_other_use_cases',1,'']]],
+  ['use_20custom_20pools_11',['When not to use custom pools',['../custom_memory_pools.html#custom_memory_pools_when_not_use',1,'']]],
+  ['user_20data_12',['Allocation names and user data',['../allocation_annotation.html',1,'index']]],
+  ['user_20data_13',['Allocation user data',['../allocation_annotation.html#allocation_user_data',1,'']]]
 ];