+++ docs/html/search/all_15.js
@@ -1,5 +1,7 @@
 var searchData=
 [
   ['warnings_0',['Validation layer warnings',['../general_considerations.html#general_considerations_validation_layer_warnings',1,'']]],
-  ['within_20budget_1',['Staying within budget',['../staying_within_budget.html',1,'index']]]
+  ['when_20not_20to_20use_20custom_20pools_1',['When not to use custom pools',['../custom_memory_pools.html#custom_memory_pools_when_not_use',1,'']]],
+  ['with_20other_20graphics_20apis_2',['Interop with other graphics APIs',['../other_api_interop.html',1,'index']]],
+  ['within_20budget_3',['Staying within budget',['../staying_within_budget.html',1,'index']]]
 ];