+++ docs/html/search/all_4.js
@@ -1,7 +1,9 @@
 var searchData=
 [
   ['enabling_20buffer_20device_20address_0',['Enabling buffer device address',['../enabling_buffer_device_address.html',1,'index']]],
-  ['explicit_20memory_20types_1',['Explicit memory types',['../choosing_memory_type.html#choosing_memory_type_explicit_memory_types',1,'']]],
-  ['exporting_20memory_2',['Exporting memory',['../opengl_interop.html#opengl_interop_exporting_memory',1,'']]],
-  ['extended_20allocation_20information_3',['Extended allocation information',['../opengl_interop.html#opengl_interop_extended_allocation_information',1,'']]]
+  ['enabling_20extensions_1',['Enabling extensions',['../quick_start.html#quick_start_initialization_enabling_extensions',1,'']]],
+  ['explicit_20memory_20types_2',['Explicit memory types',['../choosing_memory_type.html#choosing_memory_type_explicit_memory_types',1,'']]],
+  ['exporting_20memory_3',['Exporting memory',['../other_api_interop.html#opengl_interop_exporting_memory',1,'']]],
+  ['extended_20allocation_20information_4',['Extended allocation information',['../other_api_interop.html#opengl_interop_extended_allocation_information',1,'']]],
+  ['extensions_5',['Enabling extensions',['../quick_start.html#quick_start_initialization_enabling_extensions',1,'']]]
 ];