+++ docs/html/search/pages_0.js
@@ -3,6 +3,8 @@ var searchData=
   ['address_0',['Enabling buffer device address',['../enabling_buffer_device_address.html',1,'index']]],
   ['aliasing_20overlap_1',['Resource aliasing (overlap)',['../resource_aliasing.html',1,'index']]],
   ['allocation_20names_20and_20user_20data_2',['Allocation names and user data',['../allocation_annotation.html',1,'index']]],
-  ['allocator_3',['allocator',['../virtual_allocator.html',1,'Virtual allocator'],['../index.html',1,'Vulkan Memory Allocator']]],
-  ['and_20user_20data_4',['Allocation names and user data',['../allocation_annotation.html',1,'index']]]
+  ['allocator_3',['Virtual allocator',['../virtual_allocator.html',1,'index']]],
+  ['allocator_4',['Vulkan Memory Allocator',['../index.html',1,'']]],
+  ['and_20user_20data_5',['Allocation names and user data',['../allocation_annotation.html',1,'index']]],
+  ['apis_6',['Interop with other graphics APIs',['../other_api_interop.html',1,'index']]]
 ];