+++ docs/html/search/pages_12.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['within_20budget_0',['Staying within budget',['../staying_within_budget.html',1,'index']]]
+  ['with_20other_20graphics_20apis_0',['Interop with other graphics APIs',['../other_api_interop.html',1,'index']]],
+  ['within_20budget_1',['Staying within budget',['../staying_within_budget.html',1,'index']]]
 ];