+++ docs/html/search/pages_6.js
@@ -1,5 +1,5 @@
 var searchData=
 [
   ['incorrect_20memory_20usage_0',['Debugging incorrect memory usage',['../debugging_memory_usage.html',1,'index']]],
-  ['interop_1',['OpenGL Interop',['../opengl_interop.html',1,'index']]]
+  ['interop_20with_20other_20graphics_20apis_1',['Interop with other graphics APIs',['../other_api_interop.html',1,'index']]]
 ];