+++ docs/html/search/pages_a.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['opengl_20interop_0',['OpenGL Interop',['../opengl_interop.html',1,'index']]],
+  ['other_20graphics_20apis_0',['Interop with other graphics APIs',['../other_api_interop.html',1,'index']]],
   ['overlap_1',['Resource aliasing (overlap)',['../resource_aliasing.html',1,'index']]]
 ];